Pull out MC-error assertion function as fixture
